Add documentation to all AdminOrderPage functions
- Move refining of API orders up to the container level
  for AdminOrderPage.

This looks decently fine and readable; as much functionality
as possible was pulled up to the container level, but the
scratchNote and checkbox need to be kept at the level of AdminOrderItem.
